refaktoring Plate Holes float[,] na Point[]

+++ Test Projects/TwoOpt/WindowRunSalesman.xaml.cs
@@ -12,7 +12,7 @@ namespace TwoOpt
     /// </summary>
     public partial class WindowRunSalesman : IMainWindow
     {
-        public WindowRunSalesman(List<Point> Points, float fHeightToWidthRatio = 1)
+        public WindowRunSalesman(List<Point> Points, double fHeightToWidthRatio = 1)
         {
             InitializeComponent();
